Added computation of the DGWF shadowing factor.
Making modifications to skycover program using boost package to
compute polygon intersections.  Represent the focal surface perimeter
as a 100 point regular polygon.

I modified the agwsprobe.py script to produce a new set of “shadow”
files that use the 77mm diameter shadow size.  We assume that the
shadow extends symmetrically below the probe. This is not actually
true, but I don’t expect that it will significantly affect the
results.  However this should be handled correctly in any GMTO
operational software.  Also modified is the agwsvalid.cpp program and
a new shadow.cpp that computes the intersection between the probe
shadow and the DGWF field.  The numbers coming out of this seem to
make sense.  The next step is to incorporate these changes into
skycov.cpp

Shortcuts I have taken: The plate scale is hardwired in several places
  to 1”/mm.  This needs some care to sort out so we can use the
  correct dgwf value which differs by 5%. (1.05mm/asec)